evol(client): moved audio into the AudioHook
